Lab 17

+++ HW10/src/edu/miracosta/cs113/hw10/FileMergeSort.java
@@ -1,5 +1,4 @@
 package edu.miracosta.cs113.hw10 ;
-
 import java.io.File ;
 import java.io.FileNotFoundException ;
 import java.io.FileOutputStream ;
@@ -39,17 +38,17 @@ public class FileMergeSort
 	 * @param args Command Lines.
 	 */
 	public static void main(String[] args)
-	{
+	{		
 		// Files to be used
-		File originalFile = new File("OriginalFileForMergeSort.txt") ;
+		File originalFile = new File("OriginalFileForMergeSort.txt") ;	
 		File leftFile = new File("leftFile.txt") ;
 		File rightFile = new File("rightFile.txt") ;
 		File tempFile = new File("tempFile.txt") ;
 		
 		// flag
 		boolean isSorted = false ;
-		while (!isSorted )
-		{System.out.println("111") ;
+		while (!isSorted)
+		{
 			try
 			{
 				// Delete any existent file
@@ -70,6 +69,7 @@ public class FileMergeSort
 				int countSwaps = 0 ;// counts swaps if current > next
 				
 				// Separating values smaller to leftFile and larger rifgtFile
+				Integer prev = null ;
 				while (scanOriginal.hasNextInt())
 				{
 					int current = scanOriginal.nextInt() ;
@@ -81,25 +81,21 @@ public class FileMergeSort
 						leftWriter.println((next < current) ? next : current) ;
 						rightWriter.println((next > current) ? next : current) ;						
 						leftWriter.flush() ;
-						rightWriter.flush() ;						
+						rightWriter.flush() ;
+						prev = next ;
 						countSwaps = current > next ? (countSwaps + 1) : countSwaps ;						
 					}
 					else
 					{
+						countSwaps = prev != null && prev > current  ? (countSwaps + 1) : countSwaps ;	
 						leftWriter.println(current) ;
 						leftWriter.flush() ;
 					}
 				}
-				System.out.println("cout " + countSwaps) ;
-				isSorted = countSwaps == 0 ;// 0 swaps means is sorted
 				
-				// break loop if there is is no swaps
-				if (isSorted)
-				{
-					return ;
-				}
-
-				boolean isMerging = true ;
+								
+				isSorted = countSwaps == 0 ;// 0 swaps means is sorted				
+				boolean isMerging = !isSorted ;
 
 				Integer nextLeft = scanLeft.hasNextInt() ? scanLeft.nextInt() : null ;
 				Integer nextRight = scanRight.hasNextInt() ? scanRight.nextInt() : null ;
@@ -109,7 +105,6 @@ public class FileMergeSort
 				{
 					if (nextLeft != null && nextRight != null)
 					{
-
 						if (nextLeft < nextRight)
 						{
 							newWriter.println(nextLeft) ;
@@ -149,16 +144,19 @@ public class FileMergeSort
 
 				scanOriginal.close() ;
 				scanLeft.close() ;
-				scanRight.close() ;
-				
+				scanRight.close() ;				
 				// Updating original File
-				originalFile.delete() ;
-				tempFile.renameTo(originalFile) ; 
+				if (!isSorted)
+				{
+					originalFile.delete() ;
+					tempFile.renameTo(originalFile) ;
+				}
 			}
 			catch (FileNotFoundException e)
 			{
 				System.out.println(e.getMessage()) ;
 			}
 		}
+		System.out.println("DONE!") ;
 	}// End of main
 }
